Added QoL features and reworked backend
Added a progress bar for the fetching of pages and added a feh execution. Also added a homepage where it is planned to be able to execute other actions. Reworked the code base, switched "type": "module" and moved the manga fetching code to a different javascript file. index.js now only includes the homepage.